+++ client/src/app/app.module.ts
@@ -27,7 +27,7 @@ import { AddUserComponent } from './users/add-user.component';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { HomeComponent } from './home/home.component';
-import { UserCardComponent } from './users/user-card.component';
+import { UserCardComponent } from './users/user-card/user-card.component';
 import { UserListComponent } from './users/user-list.component';
 import { UserProfileComponent } from './users/user-profile.component';
