+++ client/src/app/users/user-card/user-card.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { User } from './user';
+import { User } from '../user';
 
 @Component({
   selector: 'app-user-card',
